backend/requirements.txt
bcrypt
fastapi
passlib
sqlalchemy
pydantic
PyJWT
SQLAlchemy
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
# brew install ngrok